fix: resolve cascading build errors from round 2
- exclusive_callback.h: make expectedScreen const (const virtual_identity*)
- camera.cpp: add default case for job_type_class::Carving
- plan_construct.cpp: add default for ReinforcedWall, use ByType for lever link
- population.cpp: add missing includes (body_part_raw, item, unit_wound_layerst,
  unit_patient_profile_completed_jobst), stub convict_data/victim_data/needs_recovery,
  add default for occupation_type switch
- population_death.cpp: fix lambda param for owned_buildings, simplify coffin counting
- population_pets.cpp: add buildingitemst include
- stocks_detect.cpp: simplify inventory mode check (removed Hauled/Carried)
- stocks_find.cpp: cast kitchen exc_types comparison
- stocks_update.cpp: add building_stockpilest include
- blueprint.cpp: use .string().c_str() for filesystem::path

+++ stocks_find.cpp
@@ -429,9 +429,7 @@ Stocks::find_item_info Stocks::find_item_helper(stock_item::item k)
         std::set<std::tuple<df::item_type, int16_t, int16_t, int32_t>> forbidden;
         for (size_t i = 0; i < plotinfo->kitchen.item_types.size(); i++)
         {
-            // kitchen_exc_type removed as class/namespace in Steam DF
-            // TODO: determine new kitchen exclusion type enum
-            if (plotinfo->kitchen.exc_types[i] == 1) // 1 was Cook
+            if (static_cast<int>(plotinfo->kitchen.exc_types[i]) == 1)
             {
                 forbidden.insert(std::make_tuple(plotinfo->kitchen.item_types[i], plotinfo->kitchen.item_subtypes[i], plotinfo->kitchen.mat_types[i], plotinfo->kitchen.mat_indices[i]));
             }
